+++ packages/mcp-proxy/src/tools/intercepts.test.ts
