+++ packages/mcp-proxy/src/utils/registry.test.ts
